+++ examples/particles.rs
@@ -9,7 +9,7 @@ use coffee::graphics::{
     Batch, Color, Frame, Image, Point, Rectangle, Sprite, Vector, Window,
     WindowSettings,
 };
-use coffee::input::{KeyCode, KeyboardAndMouse};
+use coffee::input::{keyboard, KeyboardAndMouse};
 use coffee::load::{loading_screen::ProgressBar, Join, Task};
 use coffee::ui::{Checkbox, Column, Element, Justify, Renderer};
 use coffee::{Game, Result, Timer, UserInterface};
@@ -91,15 +91,13 @@ impl Game for Particles {
     fn interact(&mut self, input: &mut KeyboardAndMouse, window: &mut Window) {
         self.gravity_centers[0] = input.cursor_position();
 
-        for position in input.clicks() {
-            self.gravity_centers.push(*position);
-        }
+        self.gravity_centers.extend(input.left_clicks());
 
-        if input.was_key_released(KeyCode::I) {
+        if input.was_key_released(keyboard::KeyCode::I) {
             self.interpolate = !self.interpolate;
         }
 
-        if input.was_key_released(KeyCode::F) {
+        if input.was_key_released(keyboard::KeyCode::F) {
             window.toggle_fullscreen();
         }
     }