+++ src/input/keyboard.rs
@@ -1,0 +1,6 @@
+//! Listen to keyboard events.
+
+mod event;
+
+pub use crate::graphics::window::winit::VirtualKeyCode as KeyCode;
+pub use event::Event;