+++ src/input/mouse.rs
@@ -1,0 +1,6 @@
+//! Listen to mouse events.
+
+mod event;
+
+pub use crate::graphics::window::winit::MouseButton as Button;
+pub use event::Event;