+++ src/ui/widget/checkbox.rs
@@ -4,7 +4,7 @@ use std::hash::Hash;
 use crate::graphics::{
     Color, HorizontalAlignment, Point, Rectangle, VerticalAlignment,
 };
-use crate::input::{ButtonState, MouseButton};
+use crate::input::{mouse, ButtonState};
 use crate::ui::core::{
     Align, Element, Event, Hasher, Layout, MouseCursor, Node, Widget,
 };
@@ -107,10 +107,10 @@ where
         messages: &mut Vec<Message>,
     ) {
         match event {
-            Event::MouseInput {
-                button: MouseButton::Left,
+            Event::Mouse(mouse::Event::Input {
+                button: mouse::Button::Left,
                 state: ButtonState::Pressed,
-            } => {
+            }) => {
                 let mouse_over = layout
                     .children()
                     .any(|child| child.bounds().contains(cursor_position));